Zwrocenie odpowiedniej sumy cen zamowionych zakupow
Podlicza calkowita kwote do zaplaty za posredictwem funkcji calculate w
class Bill
Please enter the commit message for your changes. Lines starting

+++ accesories.h
@@ -40,9 +40,7 @@ enum
 
 namespace shop
 {
-
-
-    typedef float towar_int_t ;
+    typedef double towar_int_t ;
     class Towar                     //ogolna klasa towaru sklepu
     {
     public:
